server/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
backports.zoneinfo==0.2.1
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
cssselect2==0.7.0
Django==4.2.16
django-cors-headers==4.4.0
django-environ==0.11.2
django-storages==1.14.4
django-xhtml2pdf==0.0.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
html5lib==1.1
idna==3.10
lxml==5.3.0
oscrypto==1.3.0
pillow==10.4.0
psycopg2-binary==2.9.10
pycparser==2.22
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.9.0
pypdf==5.1.0
pypng==0.20220715.0
python-bidi==0.6.3
python-dotenv==1.0.1
PyYAML==6.0.2
qrcode==7.4.2
reportlab==4.2.5
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
svglib==1.5.1
tinycss2==1.4.0
typing-extensions==4.12.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.3
webencodings==0.5.1
xhtml2pdf==0.2.16